Move color resources to //lib/color
Moving the files one library at a time may be easier than all at once.
