+++ lib/color/ColorPalette.cpp
