+++ lib/effect/Effect.hpp
@@ -1,10 +1,10 @@
 #ifndef __EFFECT_HPP__
 #define __EFFECT_HPP__
 
+#include <ColorPalette.hpp>
 #include <Types.hpp>
 #include <vector>
 
-#include "../../software/generic/ColorPalette.hpp"
 #include "../../software/generic/DeviceDescription.hpp"
 #include "../../software/generic/Radio.hpp"
