+++ software/generic/test/ColorPaletteTest.cpp
@@ -1,4 +1,5 @@
-#include "../ColorPalette.hpp"
+#include <ColorPalette.hpp>
+
 #include "gtest/gtest.h"
 
 TEST(ColorPalette, indexColors) {
